% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupplots.R
\name{AreaPlot}
\alias{AreaPlot}
\title{Area Plot}
\usage{
AreaPlot(
  data,
  x = "",
  y = "",
  group = "",
  color_set = "",
  title = "",
  angle = 45
)
}
\arguments{
\item{data}{frame}

\item{x}{axis name}

\item{y}{axis name}

\item{group}{column}
}
\value{
plot
}
\description{
Area Plot
}
