% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupplots.R
\name{AreaPlotGroupProportion}
\alias{AreaPlotGroupProportion}
\title{Sing-Cell Area Plot}
\usage{
AreaPlotGroupProportion(
  data,
  x = "orig.ident",
  group = "cell_type2",
  colors = "",
  title = "",
  angle = 45
)
}
\arguments{
\item{data}{frame}

\item{x}{axis name}

\item{group}{column}
}
\value{
plot
}
\description{
Sing-Cell Area Plot
}
