% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupplots.R
\name{AreaPlotProportion}
\alias{AreaPlotProportion}
\title{Sing-Cell Area Plot}
\usage{
AreaPlotProportion(
  data,
  x = "Sample",
  group = "cell_type2",
  color_set = "",
  title = "",
  angle = 45
)
}
\arguments{
\item{data}{frame}

\item{x}{axis name}

\item{group}{column}
}
\value{
plot
}
\description{
Sing-Cell Area Plot
}
