% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{BarPlot2d}
\alias{BarPlot2d}
\title{Bar plot 2d}
\usage{
BarPlot2d(
  data = NULL,
  x = NULL,
  y = NULL,
  lsz = 0.3,
  text_size = 6,
  title_size = 7,
  color = "steelblue",
  x_lab = "PLAGL2",
  y_lab = "",
  title = "",
  colors = "steelblue"
)
}
\arguments{
\item{data}{data.frame}

\item{x}{column1}

\item{colors}{color code}
}
\value{
plot
}
\description{
Bar plot 2d
}
