% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{BarPlotCount}
\alias{BarPlotCount}
\title{Bar plot count}
\usage{
BarPlotCount(data = NULL, x = "", y_lab = "", color = "#681989")
}
\arguments{
\item{data}{frame}

\item{x}{axis name}

\item{y_lab}{name}

\item{color}{color code}
}
\value{
plot
}
\description{
Bar plot count
}
