% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupplots.R
\name{BarPlotGroup}
\alias{BarPlotGroup}
\title{Bar plot with group}
\usage{
BarPlotGroup(
  data = NULL,
  group = "",
  x = "",
  y = "",
  title = "",
  x_lab = "",
  y_lab = "",
  color_set = "",
  hline = "",
  hcols = "black"
)
}
\arguments{
\item{data}{frame}

\item{group}{column}

\item{x}{axis name}

\item{y}{axis name}

\item{title}{name}

\item{x_lab}{name}

\item{y_lab}{name}

\item{color_set}{code}

\item{hline}{number}

\item{hcols}{hline color}
}
\value{
plot
}
\description{
Bar plot with group
}
