% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupplots.R
\name{BarPlotGroupPosNeg}
\alias{BarPlotGroupPosNeg}
\title{BarPlot Two Group +- in Y-axis}
\usage{
BarPlotGroupPosNeg(
  data = NULL,
  x = "",
  y = "",
  title = "",
  x_lab = "",
  y_lab = "",
  split_group = NULL,
  reverse = FALSE,
  colors = c(pos = "#FC766AFF", neg = "#5B84B1FF")
)
}
\arguments{
\item{data}{frame}

\item{x}{axis name}

\item{y}{axis name}

\item{title}{name}

\item{x_lab}{name}

\item{y_lab}{name}

\item{split_group}{column}

\item{colors}{code}
}
\value{
plot
}
\description{
BarPlot Two Group +- in Y-axis
}
