% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{BarPlotGroupText}
\alias{BarPlotGroupText}
\title{Bar plot with group & label}
\usage{
BarPlotGroupText(
  data = NULL,
  x = NULL,
  y = NULL,
  group = NULL,
  title = "",
  x_lab = "",
  y_lab = "Frequency (\%)",
  colors = c(Yes = "#D90000", No = "#DEDEDE"),
  base_lz = 0.3,
  tz = 8,
  legend_key_z = 4,
  show_text = TRUE,
  label_z = 3,
  label_suffix = "\%"
)
}
\arguments{
\item{data}{frame}

\item{x}{axis name}

\item{y}{axis name}

\item{group}{column}

\item{title}{name}

\item{x_lab}{name}

\item{y_lab}{name}

\item{colors}{code}

\item{hline}{number}

\item{hcol}{hline color}
}
\value{
plot
}
\description{
Bar plot with group & label
}
