% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{BarPlotGroupX1}
\alias{BarPlotGroupX1}
\title{Bar plot with group & label}
\usage{
BarPlotGroupX1(
  data = NULL,
  x = NULL,
  y = NULL,
  group = NULL,
  levels = NULL,
  colors = c("#E41A1C", "#377EB8"),
  title = "",
  x_lab = "",
  y_lab = "Tumor cell ratio (\%)",
  bw = 0.7,
  bls = 0.3,
  text_size = 8,
  title_size = 10
)
}
\arguments{
\item{data}{frame}

\item{x}{axis name}

\item{y}{axis name}

\item{group}{column}

\item{colors}{code}

\item{title}{name}

\item{x_lab}{name}

\item{y_lab}{name}
}
\value{
plot
}
\description{
Bar plot with group & label
}
