% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{BarPlotMeanSignal}
\alias{BarPlotMeanSignal}
\title{Bar plot for mean signal}
\usage{
BarPlotMeanSignal(
  data = NULL,
  group = "clusters",
  signal = "signal_score1",
  sort = TRUE,
  width = 0.5,
  lsz = 0.3,
  highlight = "Tumor",
  title = "",
  x_lab = "",
  y_lab = "Mean Signal"
)
}
\arguments{
\item{data}{frame}
}
\value{
plot
}
\description{
Bar plot for mean signal
}
