% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{BarPlotRatio}
\alias{BarPlotRatio}
\title{Bar plot ratio}
\usage{
BarPlotRatio(
  vec = NULL,
  title = "",
  y_lab = "Proportion (\%)",
  breaks = c("Yes", "No"),
  colors = c(Yes = "#D90000", No = "#DEDEDE"),
  sort = FALSE,
  nolegend = FALSE,
  legend_nrow = 1,
  legend_position = "bottom",
  lsz = 2,
  text_size = 5,
  title_size = 6,
  angle = 45,
  line_size = 0.3
)
}
\arguments{
\item{vec}{frame}

\item{y_lab}{name}

\item{color}{color code}
}
\value{
plot
}
\description{
Bar plot ratio
}
