% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{BarPlotSignal}
\alias{BarPlotSignal}
\title{Bar plot for signal}
\usage{
BarPlotSignal(
  data,
  x = "",
  y = "ratio",
  clusters = "clusters",
  signal = "group",
  colors = c(High = "#CB4335", Middle = "#2E86C1", Low = "#D7DBDD"),
  breaks = c("High", "Middle", "Low"),
  title = "",
  x_lab = "",
  y_lab = "The number of cells"
)
}
\arguments{
\item{data}{frame}
}
\value{
plot
}
\description{
Bar plot for signal
}
