% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{BarPlotSplitGroup_v2}
\alias{BarPlotSplitGroup_v2}
\title{Bar plot with group for positive and negative}
\usage{
BarPlotSplitGroup_v2(
  data = NULL,
  x = "",
  y = "",
  group = "",
  split_group = "",
  title = "",
  x_lab = "",
  y_lab = "",
  colors = ""
)
}
\arguments{
\item{data}{frame}

\item{x}{axis name}

\item{y}{axis name}

\item{split_group}{column}

\item{title}{name}

\item{x_lab}{name}

\item{y_lab}{name}

\item{colors}{code}
}
\value{
plot
}
\description{
Bar plot with group for positive and negative
}
