% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupplots.R
\name{BarPlotTwoGroupUD}
\alias{BarPlotTwoGroupUD}
\title{BarPlot Two Group +- in Y-axis}
\usage{
BarPlotTwoGroupUD(
  data = NULL,
  x = "",
  y = "",
  title = "",
  x_lab = "",
  y_lab = "",
  split_group = NULL,
  reverse = FALSE,
  color_set = c(pos = "#FC766AFF", neg = "#5B84B1FF")
)
}
\arguments{
\item{data}{frame}

\item{x}{axis name}

\item{y}{axis name}

\item{title}{name}

\item{x_lab}{name}

\item{y_lab}{name}

\item{split_group}{column}

\item{color_set}{code}
}
\value{
plot
}
\description{
BarPlot Two Group +- in Y-axis
}
