% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{CalBounds}
\alias{CalBounds}
\title{Calculate min and max plus function}
\usage{
CalBounds(data, minc = 0.25, maxc = 0.75)
}
\arguments{
\item{data}{value}

\item{min}{value}

\item{max}{value}
}
\value{
min and max values
}
\description{
Calculate min and max plus function
}
