% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{CallProportion}
\alias{CallProportion}
\title{Call Proportion}
\usage{
CallProportion(data, x = "", group = "")
}
\arguments{
\item{data}{frame}

\item{x}{sample name}

\item{group}{cell type}
}
\value{
data frame
}
\description{
Call Proportion
}
