% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{CallRatio_byGroup}
\alias{CallRatio_byGroup}
\title{Call Proportion by group}
\usage{
CallRatio_byGroup(data, x = "", group = "")
}
\arguments{
\item{data}{frame}

\item{x}{sample name}

\item{group}{cell type}
}
\value{
data frame
}
\description{
Call Proportion by group
}
