% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{CallRatio_byGroup2}
\alias{CallRatio_byGroup2}
\title{Call Proportion by group 2}
\usage{
CallRatio_byGroup2(data, x = "", g1 = "", g2 = "")
}
\arguments{
\item{data}{frame}

\item{x}{main column}

\item{g1}{group1}

\item{g2}{group2}
}
\value{
data frame
}
\description{
Call Proportion by group 2
}
