% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{CalMinMaxBoxplot}
\alias{CalMinMaxBoxplot}
\title{Calculate min and max for boxplot}
\usage{
CalMinMaxBoxplot(data)
}
\arguments{
\item{data}{value}
}
\value{
min and max values
}
\description{
Calculate min and max for boxplot
}
