% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{ComplexHeatmap_Basic}
\alias{ComplexHeatmap_Basic}
\title{ComplexHeatmap Simple version}
\usage{
ComplexHeatmap_Basic(
  d_mtx = NULL,
  meta = NULL,
  row_info = NULL,
  group = "cell_type2",
  sample = NULL,
  scaled = TRUE,
  col_group = NULL,
  col_sample = NULL,
  cluster_rows = TRUE,
  cluster_col = TRUE,
  show_row_dend = TRUE,
  font_size = 6,
  gap = 0.2,
  color_set = "peak",
  row_title = NULL,
  show_column_dend = FALSE,
  ht_title = "Row Z-Score",
  ht_lgd_direc = "horizontal"
)
}
\arguments{
\item{d_mtx}{dataframe}
}
\description{
ComplexHeatmap Simple version
}
