% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{ComplexHeatmap_CellType}
\alias{ComplexHeatmap_CellType}
\title{ComplexHeatmap for Cell Type as group}
\usage{
ComplexHeatmap_CellType(
  data = NULL,
  top_anno1 = NULL,
  top_anno2 = NULL,
  colors = NULL,
  col_celltype = NULL,
  col_group = NULL,
  limit_val = NULL,
  title = "Expression",
  zscore = FALSE,
  zcolors = NULL,
  show_row_names = FALSE,
  show_row_dend = FALSE,
  cluster_rows = FALSE,
  show_column_names = FALSE,
  show_column_dend = TRUE,
  cluster_columns = TRUE,
  top_annotation = TRUE,
  show_feature = NULL,
  use_raster = FALSE,
  row_split = NULL,
  font_size = 6,
  label_font = 5
)
}
\arguments{
\item{data}{dataframe}

\item{top_anno1}{top annotation group-1}

\item{top_anno2}{top annotation group-2}

\item{colors}{color set}

\item{col_celltype}{color group-1}

\item{col_group}{color group-2}

\item{zscore}{zscore yes/no}

\item{show_row_names}{yes/no}

\item{show_column_dend}{yes/no}

\item{top_annotation}{show top annotation yes/no}

\item{show_feature}{show specific gene name or not}

\item{row_split}{split row}
}
\description{
ComplexHeatmap for Cell Type as group
}
