% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R, R/heatmap.R
\name{ComplexHeatmap_Cluster}
\alias{ComplexHeatmap_Cluster}
\title{ComplexHeatmap Cluster}
\usage{
ComplexHeatmap_Cluster(data)

ComplexHeatmap_Cluster(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
ComplexHeatmap Cluster

ComplexHeatmap Cluster
}
