% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R, R/heatmap.R
\name{ComplexHeatmap_GeneExp}
\alias{ComplexHeatmap_GeneExp}
\title{ComplexHeatmap_GeneExp}
\usage{
ComplexHeatmap_GeneExp(data)

ComplexHeatmap_GeneExp(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
ComplexHeatmap_GeneExp

ComplexHeatmap_GeneExp
}
