% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{ComplexHeatmap_Group}
\alias{ComplexHeatmap_Group}
\title{ComplexHeatmap Motif-TF Groupx}
\usage{
ComplexHeatmap_Group(
  df = NULL,
  meta = NULL,
  group = "cell_type2",
  marker_info = NULL,
  levels = NULL,
  scaled = TRUE,
  col_group = NULL,
  labels = NULL,
  cluster_rows = TRUE,
  show_row_dend = TRUE,
  gap = 0.2,
  border = FALSE,
  color_set = "motif",
  ht_title = "Row Z-Score",
  row_title = NULL,
  show_column_dend = FALSE,
  legend_title = "",
  outdir = "ht_temp"
)
}
\arguments{
\item{df}{dataframe}
}
\description{
ComplexHeatmap Motif-TF Groupx
}
