% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{ComplexHeatmap_Group2}
\alias{ComplexHeatmap_Group2}
\title{ComplexHeatmap GeneExp Group Default}
\usage{
ComplexHeatmap_Group2(
  data = NULL,
  group = "",
  sample_id = "",
  colors = c("blue", "white", "red"),
  zscore = FALSE,
  row_split = NULL
)

ComplexHeatmap_Group2(
  data = NULL,
  group = "",
  sample_id = "",
  colors = c("blue", "white", "red"),
  zscore = FALSE,
  row_split = NULL
)
}
\arguments{
\item{data}{dataframe}

\item{group}{group}

\item{colors}{color set}
}
\description{
ComplexHeatmap GeneExp Group Default

ComplexHeatmap for group
}
