% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{ComplexHeatmap_GroupX}
\alias{ComplexHeatmap_GroupX}
\title{ComplexHeatmap Motif-TF Groupx}
\usage{
ComplexHeatmap_GroupX(
  df = NULL,
  meta = NULL,
  group = "cell_type2",
  sample = NULL,
  marker_info = NULL,
  filter = TRUE,
  topn = 10000,
  pct1 = "pct.1",
  avg_diff = "avg_log2FC",
  tf_gene = FALSE,
  levels = NULL,
  scaled = TRUE,
  col_group = NULL,
  col_sample = NULL,
  labels = NULL,
  cluster_rows = TRUE,
  cluster_col = TRUE,
  show_row_dend = FALSE,
  font_size = 6,
  gap = 0.2,
  border = FALSE,
  color_set = "motif",
  row_title = NULL,
  show_column_dend = FALSE,
  ht_title = "Row Z-Score",
  ht_lgd_direc = "horizontal",
  show_lgd = FALSE,
  outdir = "ht_temp"
)
}
\arguments{
\item{df}{dataframe}
}
\description{
ComplexHeatmap Motif-TF Groupx
}
