% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{ComplexHeatmap_LegendTwoGroups}
\alias{ComplexHeatmap_LegendTwoGroups}
\title{ComplexHeatmap Motif-TF Group with Legends}
\usage{
ComplexHeatmap_LegendTwoGroups(
  df = NULL,
  meta = NULL,
  group = "cell_type2",
  sample = "orig.ident",
  marker_info = NULL,
  tf_tag = FALSE,
  levels = NULL,
  scaled = TRUE,
  col_group = NULL,
  col_sample = NULL,
  labels = NULL,
  cluster_rows = TRUE,
  cluster_col = TRUE,
  show_row_dend = TRUE,
  gap = 0.2,
  border = FALSE,
  color_set = "motif",
  ht_title = "Row Z-Score",
  row_title = NULL,
  show_column_dend = FALSE,
  legend_title = "",
  ht_lgd_direc = "horizontal",
  outdir = "ht_temp"
)
}
\arguments{
\item{df}{dataframe}
}
\description{
ComplexHeatmap Motif-TF Group with Legends
}
