% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R, R/heatmap.R
\name{ComplexHeatmap_Motif}
\alias{ComplexHeatmap_Motif}
\title{ComplexHeatmap Motif}
\usage{
ComplexHeatmap_Motif(data)

ComplexHeatmap_Motif(data)
}
\arguments{
\item{data}{dataframe}
}
\description{
ComplexHeatmap Motif

ComplexHeatmap Motif
}
