% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R, R/heatmap.R
\name{ComplexHeatmap_Rank}
\alias{ComplexHeatmap_Rank}
\title{ComplexHeatmap Rank}
\usage{
ComplexHeatmap_Rank()

ComplexHeatmap_Rank()
}
\arguments{
\item{data}{dataframe}
}
\description{
ComplexHeatmap Rank

ComplexHeatmap Rank
}
