% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{ConvertWideToLong}
\alias{ConvertWideToLong}
\title{ConvertWideToLong}
\usage{
ConvertWideToLong(data, id = "", var_name = "")
}
\arguments{
\item{data}{value}

\item{id}{id.vars}

\item{var_name}{variable.name}
}
\value{
data frame
}
\description{
ConvertWideToLong
}
