% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{CorrectInfoByMarkerData}
\alias{CorrectInfoByMarkerData}
\title{Correct data for heatmap}
\usage{
CorrectInfoByMarkerData(diff_marker = NULL, mtx = NULL, meta = NULL)
}
\description{
Correct data for heatmap
}
