% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{Customize_LegendParam}
\alias{Customize_LegendParam}
\title{Customized Legend}
\usage{
Customize_LegendParam(
  title = NULL,
  col = NULL,
  nrow = 3,
  width = 2,
  height = 2,
  title_font = 6,
  label_font = 5
)
}
\arguments{
\item{title}{title}

\item{col}{color set}

\item{nrow}{show lengend row}

\item{width}{legend width}

\item{height}{legend height}

\item{title_font}{title font size}

\item{label_font}{label font size}
}
\description{
Customized Legend
}
