% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{DensityPlot_Pseudotime}
\alias{DensityPlot_Pseudotime}
\title{Density plot for pseudotime}
\usage{
DensityPlot_Pseudotime(
  data,
  x = "pseudotime",
  y = "cell_type2",
  t1 = 3,
  t2 = 15
)
}
\arguments{
\item{data}{frame}

\item{y}{axis name}

\item{x1}{axis name}

\item{group}{column}
}
\value{
plot
}
\description{
Density plot for pseudotime
}
