% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupplots.R
\name{DensityRidgesGradient_SplitGroup}
\alias{DensityRidgesGradient_SplitGroup}
\title{Density Ridges Gradient Split Group}
\usage{
DensityRidgesGradient_SplitGroup(
  data,
  x1 = "",
  y = "",
  split_group = "",
  n_row = 1
)
}
\arguments{
\item{data}{frame}

\item{x1}{axis name}

\item{y}{axis name}

\item{group}{column}
}
\value{
plot
}
\description{
Density Ridges Gradient Split Group
}
