% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupplots.R
\name{DotLinePlotGroup}
\alias{DotLinePlotGroup}
\title{Dot plot with group}
\usage{
DotLinePlotGroup(
  data = NULL,
  group = "",
  x = "",
  y = "",
  title = "",
  x_lab = "",
  y_lab = "",
  color_set = "",
  hline = "",
  hcols = "black"
)
}
\arguments{
\item{data}{frame}

\item{group}{column}

\item{x}{axis name}

\item{y}{axis name}

\item{title}{name}

\item{x_lab}{name}

\item{y_lab}{name}

\item{color_set}{code}

\item{hline}{number}

\item{hcols}{hline color}
}
\value{
plot
}
\description{
Dot plot with group
}
