% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{FilterMarkersByGroup}
\alias{FilterMarkersByGroup}
\title{Filter markers}
\usage{
FilterMarkersByGroup(
  data = NULL,
  pct1 = "pct.1",
  avg_diff = "avg_diff",
  group = "cluster",
  gene = "gene",
  n = 200
)
}
\description{
Filter markers
}
