% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupplots.R
\name{GGboxplotPval}
\alias{GGboxplotPval}
\title{ggboxplot with P-val for group}
\usage{
GGboxplotPval(
  data = NULL,
  x = "sample",
  y = "mean_peak_sig",
  group = NA,
  y_lab = "Mean peak signal",
  colors = "supp"
)
}
\arguments{
\item{data}{frame}

\item{x}{sample name}

\item{y}{value}

\item{group}{cell type}

\item{y_lab}{y lab}

\item{colors}{code}
}
\value{
plot
}
\description{
ggboxplot with P-val for group
}
