% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{Heatmap_DiffMarkers_Vert}
\alias{Heatmap_DiffMarkers_Vert}
\title{HeatMap_ExpGroups_Vertical}
\usage{
Heatmap_DiffMarkers_Vert(
  data = NULL,
  meta = NULL,
  marker_info = NULL,
  group = "cluster",
  sort_group = NULL,
  col_group = NULL,
  show_gene = NULL,
  scaled = TRUE,
  row_title = "Differentially expressed genes",
  show_column_dend = FALSE,
  legend_title = "Expression\\nZ-score",
  gap = 0.4
)
}
\arguments{
\item{data}{dataframe}

\item{meta}{top annotation group-1}

\item{marker_info}{top annotation group-2}

\item{group}{group name}

\item{sort_group}{sorted group name}

\item{col_group}{color group}

\item{show_gene}{show genes}

\item{scaled}{scaled or not}
}
\description{
HeatMap_ExpGroups_Vertical
}
