% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{Heatmap_DiffMarkers_Vertical}
\alias{Heatmap_DiffMarkers_Vertical}
\title{HeatMap_ExpGroups_Vertical}
\usage{
Heatmap_DiffMarkers_Vertical(
  data = NULL,
  meta = NULL,
  diff_marker = NULL,
  group = "cluster",
  sort_group = NULL,
  col_group = NULL,
  show_gene = NULL,
  scaled = TRUE
)
}
\arguments{
\item{data}{dataframe}

\item{meta}{top annotation group-1}

\item{diff_marker}{top annotation group-2}

\item{group}{group name}

\item{sort_group}{sorted group name}

\item{col_group}{color group}

\item{show_gene}{show genes}

\item{scaled}{scaled or not}
}
\description{
HeatMap_ExpGroups_Vertical
}
