% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{Heatmap_Motif_Group2}
\alias{Heatmap_Motif_Group2}
\title{ComplexHeatmap Motif-TF Group-2x}
\usage{
Heatmap_Motif_Group2(
  data = NULL,
  meta = NULL,
  group = "",
  sort_group = NULL,
  sample_id = "",
  diff_marker = NULL,
  show_labels = NULL,
  ht_title = "Motif Score",
  col_group = NULL,
  col_sample = NULL,
  gap = 0.4,
  border = FALSE
)
}
\arguments{
\item{data}{dataframe}
}
\description{
ComplexHeatmap Motif-TF Group-2x
}
