% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{HeatMap_Peaks_Groups}
\alias{HeatMap_Peaks_Groups}
\title{ComplexHeatmap Peak Group}
\usage{
HeatMap_Peaks_Groups(
  data = NULL,
  metadata = NULL,
  top_anno_info = NULL,
  outfile = "heatmap.peak.pdf"
)
}
\arguments{
\item{data}{dataframe}
}
\description{
ComplexHeatmap Peak Group
}
