% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{HeatmapMotif_Group2}
\alias{HeatmapMotif_Group2}
\title{ComplexHeatmap Motif Group-2}
\usage{
HeatmapMotif_Group2(
  data = NULL,
  meta = NULL,
  group = "cell_type2",
  marker_info = NULL,
  levels = NULL,
  scaled = TRUE,
  col_group = NULL,
  col_sample = NULL,
  labels = NULL,
  gap = 0.2,
  border = FALSE,
  ht_title = "Row Z-Score"
)
}
\arguments{
\item{data}{dataframe}
}
\description{
ComplexHeatmap Motif Group-2
}
