% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{HeatmapMotif_GroupX}
\alias{HeatmapMotif_GroupX}
\title{ComplexHeatmap Motif-TF Group-2x}
\usage{
HeatmapMotif_GroupX(
  data = NULL,
  meta = NULL,
  group = "cell_type2",
  marker_info = NULL,
  levels = NULL,
  scaled = TRUE,
  col_group = NULL,
  labels = NULL,
  gap = 0.2,
  border = FALSE,
  ht_title = "Row Z-Score"
)
}
\arguments{
\item{data}{dataframe}
}
\description{
ComplexHeatmap Motif-TF Group-2x
}
