% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.R
\name{Scatter2DSigSplit}
\alias{Scatter2DSigSplit}
\title{Scatter2DSigSplit}
\usage{
Scatter2DSigSplit(
  data = NULL,
  title = "",
  x = "fus_score1",
  y = "cc_score1",
  group = "group",
  x_lab = "ZR-Fusion Target Gene signal",
  y_lab = "Cell-Cycling signal",
  pt_size = 0.5,
  text_size = 5,
  split_by = "cell_type2",
  g1 = "YES",
  g2 = "NO",
  g1c = "red",
  g2c = "gray",
  hlineL = 0,
  hlineH = 0.2,
  vlineL = 0,
  vlineH = 0.2,
  lws = 0.3,
  xlim = c(-0.4, 0.5)
)
}
\arguments{
\item{data}{data frame}

\item{x}{axis x}

\item{y}{axis y}

\item{group}{group}

\item{x_lab}{x lab}

\item{y_lab}{y lab}

\item{split_by}{split by}

\item{point_size}{point size}
}
\value{
plot
}
\description{
Scatter2DSigSplit
}
