% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.R
\name{ScatterPlotHighlight}
\alias{ScatterPlotHighlight}
\title{Scatter Plot}
\usage{
ScatterPlotHighlight(
  df = NULL,
  highlight = "",
  split_by = "",
  x = NULL,
  y = NULL,
  x_lab = NULL,
  y_lab = NULL
)
}
\arguments{
\item{data}{frame}
}
\value{
plot
}
\description{
Scatter Plot
}
