% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.R
\name{ScatterPlotPlus}
\alias{ScatterPlotPlus}
\title{Scatter Plot}
\usage{
ScatterPlotPlus(
  data = NULL,
  group = "",
  x = "",
  y = "",
  x_lab = "",
  y_lab = "",
  point_size = 0.01,
  color_set = "",
  ticks = TRUE
)
}
\arguments{
\item{data}{is data frame}

\item{group}{column}

\item{x}{axis name}

\item{y}{axis name}

\item{x_lab}{name}

\item{y_lab}{name}

\item{point_size}{value}

\item{ticks}{show or not}

\item{highlight}{value}
}
\value{
plot
}
\description{
Scatter Plot
}
