% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.R
\name{ScatterPlotSplit}
\alias{ScatterPlotSplit}
\title{Scatter Plot}
\usage{
ScatterPlotSplit(
  data = NULL,
  group = "",
  x = "",
  y = "",
  x_lab = "",
  y_lab = "",
  color_set = "#ac3282",
  point_size = 0.01,
  ticks = TRUE
)
}
\arguments{
\item{data}{frame}

\item{group}{column}

\item{x}{axis name}

\item{y}{axis name}

\item{x_lab}{name}

\item{y_lab}{name}

\item{point_size}{value}

\item{ticks}{show or not}
}
\value{
plot
}
\description{
Scatter Plot
}
