% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.R
\name{ScatterPlotWithCor}
\alias{ScatterPlotWithCor}
\title{Scatter Plot With Correlation}
\usage{
ScatterPlotWithCor(
  data = NULL,
  split_by = "",
  x = "",
  y = "",
  x_lab = "",
  y_lab = "",
  dot_size = 0.5,
  dot_col = "#2278B5",
  alpha = 0.6,
  lwd = 0.5,
  reg_line_cor = "#E14C32",
  cor_method = "pearson"
)
}
\arguments{
\item{data}{data}

\item{split_by}{name}

\item{x}{axis name}

\item{y}{axis name}

\item{x_lab}{name}

\item{y_lab}{name}
}
\value{
plot
}
\description{
Scatter Plot With Correlation
}
\concept{scatter plot with correlation and split plots by split_by}
