% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.R
\name{ScatterPlotWithCorr}
\alias{ScatterPlotWithCorr}
\title{Scatter Plot With Correlation}
\usage{
ScatterPlotWithCorr(
  data = NULL,
  split_by = "",
  x = "",
  y = "",
  x_lab = "",
  y_lab = ""
)
}
\arguments{
\item{data}{data}

\item{split_by}{name}

\item{x}{axis name}

\item{y}{axis name}

\item{x_lab}{name}

\item{y_lab}{name}
}
\value{
plot
}
\description{
Scatter Plot With Correlation
}
\concept{scatter plot with correlation and split plots by split_by}
