% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ShowTopNGenes}
\alias{ShowTopNGenes}
\title{Extract top N genes from Up/Down}
\usage{
ShowTopNGenes(
  df = NULL,
  n = 10,
  gene = "gene",
  group = "signal",
  logfc = "log2FC"
)
}
\arguments{
\item{df}{data frame}
}
\description{
Extract top N genes from Up/Down
}
