% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.R
\name{Signal_UMAPPlot}
\alias{Signal_UMAPPlot}
\title{Signal UMAPPlot Group}
\usage{
Signal_UMAPPlot(
  data = NULL,
  x = "UMAP_1",
  y = "UMAP_2",
  group = "cell_type2",
  decreasing_group = TRUE,
  colors = "viridis",
  color_direc = 1,
  show_umap_lab = FALSE,
  xa = 1.2,
  xb = 0.3,
  ya = 1.1,
  yb = 0.25
)
}
\arguments{
\item{data}{dataframe}
}
\description{
Signal UMAPPlot Group
}
