% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{StandardBoxplot}
\alias{StandardBoxplot}
\title{StandardBoxplot}
\usage{
StandardBoxplot(
  data = NULL,
  x = "sample",
  y = "mean_peak_sig",
  title = "",
  x_lab = "",
  y_lab = "Mean peak signal",
  font_size = 7,
  font_size_title = 8,
  add_mean = FALSE,
  mean_dot_shape = 20,
  mean_dot_size = 3,
  nolegend = FALSE,
  comp_group = NULL,
  test_method = "wilcox.test",
  test_label = "p.format",
  font_size_pval = 2,
  outlier = 16,
  alpha = 1,
  angle = 0,
  colors = NULL
)
}
\arguments{
\item{data}{frame}

\item{x}{sample name}

\item{y}{value}

\item{y_lab}{y lab}

\item{colors}{code}

\item{group}{cell type}
}
\value{
plot
}
\description{
StandardBoxplot
}
