% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{StandardFacetWrapBoxplot}
\alias{StandardFacetWrapBoxplot}
\title{StandardBoxplot with face_wrap}
\usage{
StandardFacetWrapBoxplot(
  data = NULL,
  x = "sample",
  y = "mean_peak_sig",
  split_by = NULL,
  ncol = 3,
  title = "",
  x_lab = "",
  y_lab = "Mean peak signal",
  font_size = 6,
  font_size_title = 8,
  lsz = 0.3,
  nolegend = FALSE,
  add_test = FALSE,
  test_method = "wilcox.test",
  font_size_pval = 2,
  outlier = 16,
  alpha = 1,
  angle = 0,
  colors = NULL
)
}
\arguments{
\item{data}{frame}

\item{x}{sample name}

\item{y}{value}

\item{y_lab}{y lab}

\item{colors}{code}

\item{group}{cell type}
}
\value{
plot
}
\description{
StandardBoxplot with face_wrap
}
