% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_plot.R
\name{UMAPSignalScore}
\alias{UMAPSignalScore}
\title{UMAP - Signal Score}
\usage{
UMAPSignalScore(
  data = NULL,
  x = "UMAP_1",
  y = "UMAP_2",
  signal = "zr_score1",
  order_dec = FALSE,
  title = "",
  legend_title = "Signal",
  set_mid = NULL,
  line_size = 0.1,
  font_size = 7,
  pt_size = 0.1,
  legend_size = 2,
  colors = c("#D7DBDD", "#2E86C1", "#CB4335")
)
}
\arguments{
\item{data}{dataframe}
}
\description{
UMAP - Signal Score
}
