% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{VolcanoPlot}
\alias{VolcanoPlot}
\title{Volcano plot DEGs}
\usage{
VolcanoPlot(
  df = NULL,
  x = "logFC",
  y = "FDR",
  group = "signal",
  gene_col = "gene_symbol",
  x_lab = "log2(Fold Change)",
  y_lab = "-log10(FDR)",
  x_cutoff = 0.585,
  pt_size = 0.3,
  alpha = 0.8,
  title = "",
  title_size = 8,
  text_size = 7,
  label = NULL,
  label_size = 2,
  axis_line_size = 0.5,
  colors = c(Up = "#CB4335", Down = "#2E86C1", NotDiff = "#D2D2D2")
)
}
\arguments{
\item{df}{dataframe}

\item{x}{axis name}

\item{y}{axis name}

\item{y_lab}{name}

\item{colors}{color code}
}
\value{
plot
}
\description{
Volcano plot DEGs
}
