% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.R
\name{VolcanoPlot2}
\alias{VolcanoPlot2}
\title{Volcano plot 2 DEGs}
\usage{
VolcanoPlot2(
  df = NULL,
  x = "avg_log2FC",
  y = "p_val_adj",
  gene_col = "gene",
  x_lab = "log2(Fold Change)",
  y_lab = "-log10(FDR)",
  title = "",
  title_size = 8,
  x_cutoff = 1,
  y_cutoff = 1.3,
  lw = 0.2,
  ptz = 1,
  alpha = 1,
  text_size = 7,
  label = NULL,
  label_size = 2,
  label_color = "#A42D2D",
  show_top_n = 0,
  sort_by = "x",
  colors = c(`0` = "#525252", `1` = "#A3A3A3", `2` = "#CB4335")
)
}
\arguments{
\item{df}{dataframe}

\item{x}{axis name}

\item{y}{axis name}

\item{y_lab}{name}

\item{colors}{color code}
}
\value{
plot
}
\description{
Volcano plot 2 DEGs
}
