% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupplots.R
\name{BarPlotGroupProportion}
\alias{BarPlotGroupProportion}
\title{Bar plot for group proportion}
\usage{
BarPlotGroupProportion(data, x = "", group = "", color_set = "")
}
\arguments{
\item{data}{frame}

\item{x}{sample name}

\item{group}{cell type}

\item{color_set}{code}
}
\value{
plot
}
\description{
Bar plot for group proportion
}
