% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupplots.R
\name{BarPlotSplitGroup_v2}
\alias{BarPlotSplitGroup_v2}
\title{Bar plot with group for positive and negative v2}
\usage{
BarPlotSplitGroup_v2(
  data = NULL,
  x = "",
  y_all = "",
  y_tar = "",
  split_group = "",
  title = "",
  x_lab = "",
  y_lab = "",
  color = "blue"
)
}
\arguments{
\item{data}{frame}

\item{x}{axis name}

\item{split_group}{column}

\item{title}{name}

\item{x_lab}{name}

\item{y_lab}{name}

\item{color}{code}

\item{y}{axis name}
}
\value{
plot
}
\description{
Bar plot with group for positive and negative v2
}
