% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{ComplexHeatmap_CellType}
\alias{ComplexHeatmap_CellType}
\title{ComplexHeatmap for Cell Type as group}
\usage{
ComplexHeatmap_CellType(
  data = NULL,
  top_anno1 = "",
  top_anno2 = "",
  colors = c("blue", "white", "red"),
  col_celltype = NULL,
  col_group = NULL,
  zscore = FALSE,
  show_row_names = FALSE,
  show_column_dend = TRUE,
  top_annotation = TRUE,
  show_feature = NULL,
  row_split = NULL
)
}
\arguments{
\item{data}{dataframe}

\item{top_anno1}{top annotation group-1}

\item{top_anno2}{top annotation group-2}

\item{colors}{color set}

\item{col_celltype}{color group-1}

\item{col_group}{color group-2}

\item{zscore}{zscore yes/no}

\item{show_row_names}{yes/no}

\item{show_column_dend}{yes/no}

\item{top_annotation}{show top annotation yes/no}

\item{show_feature}{show specific gene name or not}

\item{row_split}{split row}
}
\description{
ComplexHeatmap for Cell Type as group
}
