% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{ComplexHeatmap_Group}
\alias{ComplexHeatmap_Group}
\title{ComplexHeatmap Motif-TF Groupx}
\usage{
ComplexHeatmap_Group(
  df = NULL,
  meta = NULL,
  group = "cell_type2",
  marker_info = NULL,
  topn = 10000,
  pct1 = "pct.1",
  avg_diff = "avg_log2FC",
  tf_tag = FALSE,
  levels = NULL,
  scaled = TRUE,
  col_group = NULL,
  labels = NULL,
  cluster_rows = TRUE,
  cluster_col = TRUE,
  show_row_dend = TRUE,
  gap = 0.2,
  border = FALSE,
  color_set = "motif",
  ht_title = "Row Z-Score",
  row_title = NULL,
  show_column_dend = FALSE,
  legend_title = "",
  ht_lgd_direc = "horizontal",
  outdir = "ht_temp"
)
}
\arguments{
\item{df}{dataframe}
}
\description{
ComplexHeatmap Motif-TF Groupx
}
