% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R, R/heatmap.R
\name{ComplexHeatmap_Group2}
\alias{ComplexHeatmap_Group2}
\title{ComplexHeatmap GeneExp Group Default}
\usage{
ComplexHeatmap_Group2(
  data = NULL,
  group = "",
  sample_id = "",
  row_split = NULL
)

ComplexHeatmap_Group2(
  data = NULL,
  group = "",
  sample_id = "",
  row_split = NULL
)
}
\arguments{
\item{data}{dataframe}
}
\description{
ComplexHeatmap GeneExp Group Default

ComplexHeatmap GeneExp Group Default
}
