% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{ComplexHeatmap_Group2b}
\alias{ComplexHeatmap_Group2b}
\title{ComplexHeatmap GeneExp Group-2}
\usage{
ComplexHeatmap_Group2b(
  data = NULL,
  group = "",
  sample_id = "",
  ht_title = "",
  col_group = NULL,
  col_sample = NULL,
  zscore = FALSE,
  cluster_columns = TRUE,
  feature_info = NULL,
  gene = "gene",
  levels = NULL
)
}
\arguments{
\item{data}{dataframe}
}
\description{
ComplexHeatmap GeneExp Group-2
}
