% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R, R/heatmap.R
\name{ComplexHeatmapMotif_Group2x}
\alias{ComplexHeatmapMotif_Group2x}
\title{ComplexHeatmap Motif-TF Group-2x}
\usage{
ComplexHeatmapMotif_Group2x(
  data = NULL,
  group = "",
  sample_id = "",
  ht_title = "Motif Score",
  col_group = NULL,
  col_sample = NULL,
  zscore = FALSE,
  cluster_columns = TRUE,
  cluster_rows = FALSE,
  show_row_dend = FALSE,
  show_column_dend = TRUE,
  feature_info = NULL,
  gene = "gene",
  labels = NULL,
  levels = NULL,
  limits = NULL,
  colors = c("blue", "white", "red"),
  gap = 0.4,
  border = FALSE,
  max_cutoff = NULL
)

ComplexHeatmapMotif_Group2x(
  data = NULL,
  group = "",
  sample_id = "",
  ht_title = "Motif Score",
  col_group = NULL,
  col_sample = NULL,
  zscore = FALSE,
  cluster_columns = TRUE,
  cluster_rows = FALSE,
  show_row_dend = FALSE,
  show_column_dend = TRUE,
  feature_info = NULL,
  gene = "gene",
  labels = NULL,
  levels = NULL,
  limits = NULL,
  colors = c("blue", "white", "red"),
  gap = 0.4,
  border = FALSE,
  max_cutoff = NULL
)
}
\arguments{
\item{data}{dataframe}
}
\description{
ComplexHeatmap Motif-TF Group-2x

ComplexHeatmap Motif-TF Group-2x
}
