% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{CorrectInfoByMarkerData}
\alias{CorrectInfoByMarkerData}
\title{Correct data for heatmap}
\usage{
CorrectInfoByMarkerData(
  mtx = NULL,
  meta = NULL,
  marker_info = NULL,
  pct1 = "pct.1",
  avg_diff = "avg_diff",
  group = "cluster",
  gene = "gene",
  n = 200
)
}
\description{
Correct data for heatmap
}
