% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{DensityPlot_Pseudotime}
\alias{DensityPlot_Pseudotime}
\title{Density plot for pseudotime}
\usage{
DensityPlot_Pseudotime(
  data,
  x = "pseudotime",
  y = "cell_type2",
  t1 = NULL,
  t2 = NULL
)
}
\arguments{
\item{data}{frame}

\item{x}{axis name}

\item{y}{axis name}

\item{t1}{early time point}

\item{t2}{later time point}
}
\value{
plot
}
\description{
Density plot for pseudotime
}
