% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{HeatmapMotif_Group}
\alias{HeatmapMotif_Group}
\title{ComplexHeatmap Motif-TF Group-2x}
\usage{
HeatmapMotif_Group(
  data = NULL,
  group = "cluster",
  gene = "gene",
  sample_id = "orig.ident",
  levels = NULL,
  scaled = TRUE,
  marker_info = NULL,
  col_group = NULL,
  labels = NULL,
  gap = 0.4,
  border = FALSE,
  max_cutoff = NULL,
  ht_title = "Row Z-Score"
)
}
\arguments{
\item{data}{dataframe}
}
\description{
ComplexHeatmap Motif-TF Group-2x
}
