% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{HeatmapMotif_Group2}
\alias{HeatmapMotif_Group2}
\title{ComplexHeatmap Motif Group-2}
\usage{
HeatmapMotif_Group2(
  data = NULL,
  group = "cluster",
  gene = "gene",
  sample_id = "orig.ident",
  levels = NULL,
  scaled = TRUE,
  marker_info = NULL,
  col_group = NULL,
  col_sample = NULL,
  labels = NULL,
  gap = 0.4,
  border = FALSE,
  max_cutoff = NULL,
  ht_title = "Row Z-Score",
  show_legend = FALSE
)
}
\arguments{
\item{data}{dataframe}
}
\description{
ComplexHeatmap Motif Group-2
}
