% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.R
\name{ScatterPlotWithCorr}
\alias{ScatterPlotWithCorr}
\title{Scatter Plot With Correlation}
\usage{
ScatterPlotWithCorr(
  df = NULL,
  split_by = "",
  x = NULL,
  y = NULL,
  x_lab = NULL,
  y_lab = NULL
)
}
\arguments{
\item{data}{frame}
}
\value{
plot
}
\description{
Scatter Plot With Correlation
}
\concept{scatter plot with correlation and split plots by split_by}
