% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.R
\name{Signal_UMAPPlot}
\alias{Signal_UMAPPlot}
\title{Signal UMAPPlot Group}
\usage{
Signal_UMAPPlot(
  data = NULL,
  x = "UMAP_1",
  y = "UMAP_2",
  group = "cell_type2",
  title = "",
  decreasing_group = FALSE,
  point_size = 0.01,
  color = "rocket",
  color_direc = -1,
  color_limits = c(0.1, 3),
  show_umap_lab = FALSE,
  xa = 1.1,
  xb = 0.3,
  ya = 1.05,
  yb = 0.25,
  x_hjust = 0.03,
  y_hjust = 0.04
)
}
\arguments{
\item{data}{dataframe}
}
\description{
Signal UMAPPlot Group
}
