% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{StandardBoxplot}
\alias{StandardBoxplot}
\title{StandardBoxplot}
\usage{
StandardBoxplot(
  data = NULL,
  x = "sample",
  y = "mean_peak_sig",
  title = "",
  x_lab = "",
  y_lab = "Mean peak signal",
  font_size = 7,
  font_size_title = 8,
  nolegend = FALSE,
  add_test = FALSE,
  test_method = "wilcox.test",
  show_pval_position = 10,
  outlier = 16,
  alpha = 1,
  colors = "blue"
)
}
\arguments{
\item{data}{frame}

\item{x}{sample name}

\item{y}{value}

\item{y_lab}{y lab}

\item{colors}{code}

\item{group}{cell type}
}
\value{
plot
}
\description{
StandardBoxplot
}
