% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{StandardBoxplot}
\alias{StandardBoxplot}
\title{ggboxplot with P-val for group}
\usage{
StandardBoxplot(
  data = NULL,
  x = "sample",
  y = "mean_peak_sig",
  group = NULL,
  x_lab = "",
  y_lab = "Mean peak signal",
  font_size = 7,
  font_size_title = 8,
  nolegend = FALSE,
  add_test = FALSE,
  outlier = 16,
  colors = ""
)
}
\arguments{
\item{data}{frame}

\item{x}{sample name}

\item{y}{value}

\item{group}{cell type}

\item{y_lab}{y lab}

\item{colors}{code}
}
\value{
plot
}
\description{
ggboxplot with P-val for group
}
