% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_plot.R
\name{UMAPSignal}
\alias{UMAPSignal}
\title{UMAP - Signal}
\usage{
UMAPSignal(
  data = NULL,
  x = "UMAP_1",
  y = "UMAP_2",
  group = "group",
  signal = "zr_score1",
  order_dec = FALSE,
  title = "",
  line_size = 0.1,
  font_size = 7,
  pt_size = 0.01,
  legend_size = 2,
  legend_title = NULL,
  breaks = waiver(),
  colors = c("#CB4335", "#2E86C1", "#D7DBDD")
)
}
\arguments{
\item{data}{dataframe}
}
\description{
UMAP - Signal
}
