% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.R
\name{UMAPSignal}
\alias{UMAPSignal}
\title{UMAPSignal}
\usage{
UMAPSignal(
  data = NULL,
  x = "UMAP_1",
  y = "UMAP_2",
  signal = "pseudotime",
  title = "",
  order_dec = FALSE,
  point_size = 0.01,
  color = "rocket",
  color_direc = -1,
  show_umap_lab = FALSE
)
}
\arguments{
\item{data}{dataframe}
}
\description{
UMAPSignal
}
