% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_plot.R
\name{UMAPSignalSplit}
\alias{UMAPSignalSplit}
\title{UMAPPlot - split per sample}
\usage{
UMAPSignalSplit(
  data = NULL,
  x = "UMAP_1",
  y = "UMAP_2",
  sample = "Sample",
  signal = "signal",
  group = "group",
  sorted_sample = "",
  order_dec = FALSE,
  pt_size = 0.01,
  line_size = 0.1,
  title = "",
  legend_title = "Signal",
  legend_size = 2,
  colors = c(High = "#CB4335", Medium = "#2E86C1", Low = "#D7DBDD"),
  breaks = c("High", "Medium", "Low"),
  ncol = 6
)
}
\arguments{
\item{data}{dataframe}
}
\description{
UMAPPlot - split per sample
}
